\name{WMCapacity-package}
\alias{WMCapacity-package}
\alias{WMCapacity}
\docType{package}
\title{
Estimation of working memory capacity from change detection data
}
\description{
The WMCapacity package implements the hierarchical Bayesian multinomial
models of Morey (2011), for the estimation of working memory
capacity from change detection data.
}
\details{
\tabular{ll}{
Package: \tab WMCapacity\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2009-06-20\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
Most users will want to use this package through the
\code{\link{wommbatGUI}} function.
}
\author{
Richard D. Morey <richarddmorey@gmail.com>
}
\references{
Morey, R. D. (2011). A hierarchical Bayesian model for the measurement of working memory capacity. Journal of Mathematical Psychology, 55, 8-24
}
\keyword{nonlinear}


